Address comment, add aws util
