+++ entrypoint/upgradecluster/upgradecluster_suite_test.go
@@ -9,7 +9,6 @@ import (
 	. "github.com/onsi/gomega"
 
 	"github.com/rancher/distros-test-framework/config"
-	"github.com/rancher/distros-test-framework/pkg/aws"
 	"github.com/rancher/distros-test-framework/pkg/customflag"
 	"github.com/rancher/distros-test-framework/shared"
 )
@@ -18,7 +17,6 @@ var (
 	kubeconfig string
 	flags      *customflag.FlagConfig
 	cluster    *shared.Cluster
-	awsClient  *aws.Client
 	err        error
 )
