+++ entrypoint/upgradecluster/upgradenodereplacement_test.go
@@ -4,12 +4,10 @@ package upgradecluster
 
 import (
 	"fmt"
-	"sync"
 
 	"github.com/rancher/distros-test-framework/pkg/assert"
 	"github.com/rancher/distros-test-framework/pkg/aws"
 	"github.com/rancher/distros-test-framework/pkg/testcase"
-	"github.com/rancher/distros-test-framework/shared"
 
 	. "github.com/onsi/ginkgo/v2"
 )
@@ -85,32 +83,11 @@ var _ = Describe("Upgrade Node Replacement Test:", Ordered, func() {
 
 	AfterAll(func() {
 		if flags.Destroy {
-			deleteAWSResources()
+			aws.DeleteEC2Instances(cluster)
 		}
 	})
 
 })
-
-func deleteAWSResources() {
-	ips := shared.FetchNodeExternalIPs()
-	awsClient, err = aws.AddClient(cluster)
-	if err != nil {
-		shared.LogLevel("error", "error creating aws client: %w\n", err)
-	}
-	var wg sync.WaitGroup
-	for _, ip := range ips {
-		wg.Add(1)
-		go func(ip string) {
-			defer wg.Done()
-			nodeDelErr := awsClient.DeleteInstance(ip)
-			if nodeDelErr != nil {
-				shared.LogLevel("error", "on deleting node with ip: %v, got error %w", ip, nodeDelErr)
-				return
-			}
-		}(ip)
-	}
-	wg.Wait()
-}
 
 var _ = AfterEach(func() {
 	if CurrentSpecReport().Failed() {
